KSENROLL-1223: Fix null fields conversion issue
The TestLprServiceMockImpl file was seeing TransactionItem Results
objects.  But this is only possible after a submit not after a create.

The problem is in how the Entity and Info objects are converted in that
not enough care was taken to only add the object if the fields for the
object are defined.  i.e. if A.1 and A.2 are both null then don't create
the A object.