KSENROLL-1515 initial mocked test structure for reg groups and seat pools
This is the initial test structure for the Mocked at class 2 level
CourseOfferingService.

Still not at the Registration Group Stage yet but working to get the Class
2 data model built to support those tests.

I've copied the original test cases and now have to craft class 2 objects
to represent what they want to do vs the at class1 details that was
contained in the previous "with mock" implementation.