KSENROLL-1515 More setup for testing the class2 objects in CourseOfferingService
The CourseOfferingServiceDataUtils is almost complete and able to build
the basic Offering objects for testing things out at the class 2 object
level.

I started with the TestCourseOfferingServiceClass2MockImpl and it is
included here so that it will be committed at least once into the
repository.

But it turned out that a Mock Implementation was started with SOC part of
CourseOffering.  This is much more robust that mine so I have switched
over to it for the TestCourseOfferingServiceMockImpl class.

There is still more work to do to get the existing test cases to pass with
the mock impl and to actually get to the real issue of creating unit tests
for the Reg Groups and Seat Pools.