KSENROLL-1654 Create a reload mechanism for mock service data
For testing the mock impl we want to have a full array of data.

For class1 services this test data is easily manipulated but for class2
more support is needed to coordinate the data coming from different
class1 and class2 services.

Two new interfaces are created to:
1. Allow Mock impl's to reset back to their initial emtpy state
2. Allow the data loader to be aware of the test execution sequence to
reset after the test and reload before the test.

It is then implemented for the CourseOfferingService so that all of the
related services are cleared and the entire state is rebuilt before
executing the next test.

The main reason for this is to allow deep structural changes to be made in
tests without effecting the other tests.