KSENROLL-1515 Create Registration Code Generator and improve generate reg groups
Created an interface that can be implemented to provide a spring bean that
will generate a unique registration group code.  This is a code that will
be visible to the student when they register.

Also supports adding additional activity offerings and having the
new registration group combinations to be generated.