KSENROLL-1515 Enable Validation Decoratorfor CourseOfferingServiceMockImpl
I copied the validation decorator stuff from the co-test-context.xml
into the mock context.

This commit contains the updates to the Registration Group and
SeatPoolDefinitions to let the unit tests pass using the validation
decorator.

There was some additional wiring required to get the RegistrationGroups
and SeatPool main dictionary documents to load but it now works.

Also the CourseOfferingServiceMockImpl needed to be a top level bean in
order to appear in the list of MockServices that are cleared after each
test case is run.