KSENROLL-1515
Brings the TestCourseOfferingServiceImpl test suite to the point of almost
running.

Added a boolean option that will allow the data to be reloaded at each
test in the mock tests but only on startup for the database backed tests.
This is ok because against the database all of the transactions rollback
vs committing.

Adjusted the courseOffering-persistence.xml for recent changes to that the
tests would get closer to running.

There is some inconsistency in the test data right now but it should be
fairly close to being saveable.