KSENROLL-1515 Cascaded deletes, dependant object tests and new test cases
deleteActivityOffering is now aware of the existence of RegistrationGroups
and SeatPools.

A new deleteActivityOfferingCascaded is created to delete the AO and
related RegGroups and SeatPools.

The dependant object exists checks were added into the validation
decorator.

Added a bunch of new test cases to confirm that delete detects the
dependant object and that cascade does in fact delete the dependant
objects.