KSENROLL-1969 Add inherited methods into service method output
Incorporate methods contributed by parent interfaces that are not
themselves services into the operations and method details for the
service.

This was added to allow the CourseOfferingService to show the details for
methods that are contained on the CourseOfferingServiceBusinessLogic
interface which is not a service.

But still allows CluService to link against VersionManagementService which
is a seperate service.