KSENROLL-501 delete eclipse launchers from ks
The launchers aren't in the svn:ignore but they need to be specificially
created and saved into the repository which shouldn't happen in the
future.

The problem is that they are auto imported and eclipse will just update
the settings which breaks backwards compatibility.

If they are ever needed they can be put into a seperate repository.