KSENROLL-2326 refactor to allow mock and validation to use same base
We extract an AbstractKSMojo class that defines the <sourceDirs> parameter
aswell as the getModel and validateModel methods.

The original mojo's still work and now we have the basis for generating
the mock and validation decorators using maven to provide the source
paths.

We also fix a bug where the output stream was not being closed properly
when generating the contract docs.

And an issue for mock generation where we were trying to get a parameter
from a method that had none.  This code was not even being used removing
it is fine.