KSENROLL-3171 and KSENROLL-3182 GraphViz charts and service client.
KSENROLL-3171 creates the ks-graphviz-charts artifact that can generate
graphviz charts for course offerings.

KSENROLL-3182 creates a service client that will consume data from the
deployed test environments.

The first pass is to collect course offerings where the number of
activities are > 25.

The Main.java in the test code will use the graphviz generator to emit the
contextualized chart for each course being evaluated.