Beginning of the switchover to using two page element scripts.
The two page element documents will refer to different parts of Sakai.
Site_page_elements will contain Page Classes for pages that appear in
the context of a Course or Portfolio Site. The admin_page_elements
script will contain page classes that exist for the Admin user, in the
context of "My Workspace". Note the classes that are exclusive to the
admin_page_elements script are written to be used ONLY in the context
of "My Workspace". They will fail if used inside the context of the
"Administration Workspace."