Completion of the tc_calendar_events.rb test case, among other things. See details...
Changed how all of the attach files classes work. Created a superclass
containing all of the needed methods, with class variables being
initialized in each of the associated subclasses. May want to
eventually remove the class variables completely from the subclasses
and instead define them in a method that resides in the superclass. Not
necessary now, however.  Changes to the other scripts are all basically
tweaks and bug fixes, mostly related to the attach superclass updates.