Added some utility methods and fixed a bug in tc_create_assignments
Added 3 new methods: last_month, current_month, next_month. These
return 3-char strings equal to the relevant month. Useful for the Month
select lists.

Fixed a bug in the create assignments test case where the script would
error out on the first of the month, because the open date would be
made to be later than the due date.