KSENROLL-2868 Add support for local resolution of top level bean
In the service application context there will be multiple implementations
of the service interface.

i.e. authorization decorator -> validation decorator -> impl -> business
logic.

We want to always wire using the top object in the chain.

This commit adds in the ability to use the @WebService.serviceName (or
@WebService.name if the serviceName is not defined) to resolve the top
level bean.

This requires that the bean intended to be the top level bean has an id
matching captialized name of the java interface for the service.

For example "CourseOfferingService" is the expected top level bean id for
the implementations of the CourseOfferingService.