KSENROLL-3775 Implement in memory KRMS unit test
TestEnrollmentEligibility tests the new
CompletedCourseBetweenTermsProposition.

The test case shows how a data loader can be used in conjunction with the
MockService implementations to reload the entire service dependency data
between each test case.

This allows mock data to be inserted into the AcademicRecordService for
each test; and the KRMS execution to occur to validate the functionality
of the Propositions and Term Resolvers.

The KRMSEngineService which also implements MockService allows the term
resolvers to be specified differently for each test.

The TestKRMSEnrollmentEligibilityDataLoader extends a new base
AbstractMockSErvicesAwareDataLoader (committed seperately into
ks-common-test).

This commit also includes changes to the CourseOfferingSetService and
CourseOfferingService DataLoaders to also use this new base class.