KSENROLL-4671 Add TestAtpServiceOptimisticLocking test case
Adds in a new test case for testing optimistic locking with the
AtpServiceImpl.

Also fixes the implementation of updateAtp so that the optimistic locking
exceptions are thrown properly.

Because we need to get rid of the find I added a new
entityExists(primaryKey) method into the GenericEntityDao that can be used
to accommplish does it exist test.

I had to add an EntityManager.flush() call at the end of the update to
make sure the version was updated in the object graph returned from the
updateAtp method.