KSENROLL-4671 detach before merge in GenericEntityDao
In order for the optimistic locking exceptions to be thrown by JPA
the enity being merged can't be known to the EntityManager.

The typical update pattern is to load the existing entity and then merging
the DTO data into it before saving it back to the database.

This commit changes the behaviour of the GenericEntityDao.merge method to
call detach before calling merge.