KSENROLL-4761 Set fake id on entity update to let test pass
In AtpServiceImpl.updateAtp it is an error for the id of the entity to
update to be null.

Either the new top level aop based decorator should catch it or the
GenericEntityDao.find method should return null if the if is null.

For now the fix is to set an id value for the non-persisted entity to a
fake value. This avoids the null key update case for the unknown key
update case.