KSENROLL-4997 Improve debugging output for AttributeTester
When attribute tester fails because the expected and actual keys are
different instead of just printing out those two keys print the entire
expected and actual lists.

This helps to show where the problem is located and is faster than having
to inspect the code in a debugging session.

Set to the debug level so that it will not normally print.