KSENROLL-5326 Add override mechanism for log4j.properties in tests
Creates a special configurer that can be used to look for a system
property called 'log4j.properties' and if it exists use that value
as the absolute path to an alternate log4j.properties file
and configure log4j with it.

Adds a unit test that verifies that this override mechanism works by
changing the default log level to DEBUG and verifying that within the
test the debug log level is setup as requested.

Usage: log = KSLog4JConfigurer.getLogger(clazz);