KSENROLL-5350 Add initial methods for user story 8
These are for the total number of seats available for a RG, AOC and CO.

The total is capped by the max (ao.maxEnrollment) for each ao type.

The exact way the aggregation works at the AOC and CO level is still a
work in progress.

The unit tests should help to flush out these details.