KSENROLL-5094 Add unit working unit test for user story 3
Adds a unit test to test that using the app layer the new RG's are
created when a new AO is added.

Also includes enhancements to the CourseOfferingDataLoader to be able to
construct new course offerings with varying ao and aoc and maxEnrollment
characteristics.

The unit test for user story 8 is started but it does not yet work
properly.