KSENROLL-3907 Adding capability to acquire md5 sums for copyfrom revs for v3 streams
V3 streams store the delta of the file against its prior version.  This is
a problem for the Prepare.py script because within git this delta version
is not known so the computed md5 hash is not going to be correct.

This commit tries to let the node filter acquire these details when the
copyfrom commit is copied.

Original Git Author Date: Thu 14 Mar 2013 10:03:40 AM EDT