KSENROLL-3907 Track the copyFromPrefix properly
The target rev can be an incremental dump but the copyfrom rev needs to
have all of the files available.

Then on the java side the copyFromPrefix is needed so that we can
normalize the copyfrom rev dump into the path structure resolved in the
git analysis.

Original Git Author Date: Fri 15 Mar 2013 12:43:25 PM EDT