KSENROLL-3907 Fix inputstream.skip logic
After adding the capability to parse the revision properties I was not
skipping the file properly which was causing the parser to end up reading
line by line through the binary content.

Now the tests and the merge detector run.

There are still issues with the merge detector in that the svn:mergeinfo
was assumed to live on the copy from path nodes but that is not correct.