KSENROLL-7896 Create program to extract missing revision properties
The refactoring of the SvnDumpFilter caused a problem in that the revision
properties were not written into the filtered dump file.

Actually the content was but because of the 'prop-content-length' property
being missing svnrdump load did not load the data.

Or show any error for that matter.

Add a new program ExtractRevisionPropsMain that will read the original svn
dump file, extract the revision properties and transform them into a
script that can be called against the target repository to apply them.