KSENROLL-6924 Fix unit test path issue to let tests run
Without this fix I could not run 'mvn clean install' to build the ks-sonar
project.

The "sql/missing_FK_query.sql" is specified in the
TestDatabaseIntegrityScript unit test.

But that path exists in src/main/resources not relative to the project
root directory that the tests run from.

I used the available ClassLoader to lookup the resource and then the
commons-io IOUtils.toString method to read the file content into a string
for use in the PreparedStatement.

I also added a log4j.properties file to let the output showup when running
the tests.