KSENROLL-8049 Add show-branch.sh script and some other helpers
The comapre-revs.sh script is used when performing sync merges but it
required manually tagging the git commit graph and calculating the
differential.

The show-branch.sh script leverages the existing git show-branch and git
merge-base commands to automatically identify the branch start point and
the commits.

The data can be used to cherry pick the revisions and then to construct a
script that can be run to change the svn:author revison properties of the
replayed commits back to the original author.