KSENROLL-8320 protect ks-ap module aswell
Tested the pre-commit hook on a copy of the ks repository and noticed that
while it worked ok there was a ks-ap module that was not being protected.

This commit adds the ks-ap to the list of protected enrollment modules.

Fix the log message when trying to delete a procted contrib/CM path.

Was 'echo "" >2&' when it should have been 'echo "" >&2'