KSENROLL-8320 Autodetect structural modules to protect
Instead of listing out the modules we need to protect, we use svn itself to
tell us which modules exist in the paths we want to protect.

enrollment/$module/{branches,tags,trunk}

contrib/$module/{branches,tags,trunk} (NOT CM)

contrib/CM/$module/{branches,tags,trunk}

There is no restriction on the creation of these paths so as new ones are added
the hook will pick them up automatically.

A future enhancement could be to cache the protected paths and only update when
we detect a new module being added.

Also updates the mock repository structure to include the other contrib modules
and the enrollment/ks-ap module.