KSENROLL-9987 Set Service Name for KS Permission Types in KIM data
In order to support running in distributed with KIM and KEW embedded we
need to have a service named defined for all KS permission types.

In Bundled we had been setting up the defaultPermissionTypeService and
wiring it into the KimService but stock rice doesn't work that way.

Without this change against stock rice there are NullPointerExceptions
when the Permissions are looked up.