Replay of r55804 from contrib | nnacion | 2013-09-12 12:21:48 -0400 (Thu, 12 Sep 2013) | 5 lines
KSCM-851 Permission Details being ignored on service call

The call to the isAuthorizedByTemplate service in AbstractDataService isn't
correctly passing in the proper permission details. Instead it is passing
in an empty map. This is causing the permission details to be ignored and
is granting permissions beyond the scope of what they are configured to do.

This fix brings back the permission details setup and passes it on to be
used.