KSENROLL-11841 improve space of revision map, specifiy repo uuid.
The revision mapper has been reworked to store the revisions in a single
snappy compressed file with an index that has the byte offset and byte
length of that segment.

Add support for paths starting with 'merges' treat it like 'branches'
where the next element is the name of the branch.

Add ability to specific the repository uuid to use in the git-svn-id in
the commit messages.