KSENROLL-11841 Various Improvements
- optimize revision mapper

switch to using a random access file and to storing the index in memory.

we only source from disk on startup.

- add an revision map repack phase

In order to maximum compression we should write all of the revision
changes in oneshot.

but this is slow if done for every revision.

Instead create a repack method that can be called when the Git GC is
occuring.

This will allow the maping file to achieve maximum compression.

- Add org.kuali.student.git.importer.ConvertOldBranchesToTagsMain program
  to create tags from branches that contain @

tried out moving the tag back up the tree but it is not accurate right now.

it works but more info is needed before it should be used in production.

remove walk back and implement delete.

this works nicely.

- repack the map file every 10k revs

this doesn't scale well

- switch to bzip2 streams from snappy for the revision map

- reworked the branch detection logic into spring beans

Instead of being locked in a utility class we can now get the data using
a spring bean.

This is just the first step to transfer into the new architecture.

- added an option to skip the gc

- add support for target is branch subtree source is a branch

for directory add

- add a test case for directory copying from a branch into a subdirectory of
a current branch.

- spaces in branch names are --- and handle delete path properly

- fix bug in deleting paths

Delete path was cascading so ks-1.3 would delete ks-1.3-ec1 which was
wrong.