KSENROLL-11841 svn:mergeinfo and svn:externals related changes
- Support default branches for blob adds and changes

In some cases the normal branch processing won't find a valid branch.  But
in the case of blobs we want to store them somewhere in case they are ever
referred to in a copyfrom.

So in the blob add or change case where a branch is not naturally found we
will create one using the first element in the blob path.

This will not handle the case where there is a file added to the root of
the repository.

- Fixed handling of the contrib/CM path with invalid tags

Because it was a direct path from contrib/CM but had invalid tags it was
missclassified.  Fixed the rules so that it will be classified correctly
as a branch.

- Strip *'s from --record-only merges in the svn:mergeinfo property
  content.

- svn:mergeinfo Improvements

== Look at specific consectutive commits not revisions when determining
the max contiguos revision to merge.

== Collect the merge paths per revision and write at the end.

Fixes where svn:mergeinfo existed on a subtree of the target branch and
was overwriting the whole branch info.

Now we aggregate all the svn:mergeinfo property content for the current
revision which fixes the problem.

== Create a program to print out a specific revision from the data file.

Used to read the revision map data file at a certain revision.

- Generate sha1 based externals file for fusion-maven-plugin

Only the repo with the jsvn data would be able to fuse the old externals.

This will read the contents of the svn:externals attribute and if details
are present it will change it to a sub directory to branch and sha1.

This can be used by the fusion-maven-plugin in the future to fuse any old
revision.