KSENROLL-11841 Fix not saving delete changes issue
Traced the problem to optimization logic when the commit is created.

If there are no changes then reuse the original tree however we forgot to
include deletes as changes in the condition.

This caused the delete which was correctly stored in the GitTreeData to
not be converted into a JGitTreeData object and not persisted.