KSENROLL-11841 adapt to delta properties
It turned out that the revision properties of svn:mergeinfo and
svn:externals are only emited in a delta fashion (when the --incremental
option is used to create the svn revision 2 dump file)

This means that even if there is no mention of the property in the current
revision it should be persisted as if it did exist.

It also means that absense of this property does not mean the meta data
should be deleted.  The delete can only occur when the key exists with an
empty string value.

For svn:mergeinfo we just carry the value forward from the previous commit
if it exists.

For svn:externals we need to extract the ExternalModuleInfo from the
existing 'fusion-maven-plugin.dat' file if it exists in the parent tree.

Then we will update the branch sha1's as appropriate.

We could create a different commit to hold the fusion-maven-plugin.dat
changes when commits are made to the sub modules.

But the first pass on this will only construct an updated commit if there
is a commit on the root.