KSENROLL-11841 Fix commit date issue
The previous fix brought the times from 8 hours off to 4 hours but this
takes it the rest of the way.

We assume and it looks likely that subversion stores the dates in UTC but
we need to convert into the local timezone when we attach the date to the
commit.

Reworked how this works and used a test case from ks rev 51525 comparing
it to how it appears in the dump file and how it appears in svn log.