KSENROLL-11841 Handle no author date or message revision
KC rev 15502 has no author, date or commit message so rework the code to
handle this case.

While this rev has no commits don't make the assumption it will be like
that in all cases.

So we stub in a fake user and date and make a note in the message.

We get the time of the previous commit + 5 minutes or set to the epoch if
this is the first commit.