KSENROLL-11841 Add copyfrom file with prop change test case
Found a problem in the KS conversion where some files were showing zero
content.

The case was identified as a direct file copy combined with a property change
on that file.

Sample import r 18 replicates the problem as was used to refactor how adding
a blob works so that if the content length == prop content length check if
there is copyfrom data and if so use that for the blob source.