KSENROLL-11841 Attempt to fix parentage history loss issue
In some cases even though the parents exist for a given commit it is not being
recorded that way.

This switches from using the git ref lookup to the revision map database which
was persisted on the last revision as a way to determine the parentage of a new
commit.

We also refresh the ref database after a garbage collection.