KSENROLL-13873 Add jung based repository viewer
Add in a repository-viewer module that will construct a jung2 swing viewer of
a git repositories commit graph.

Structurally it is looking correct and mouse zooming in and out works.

the graph is presented in reverse with the most recent commits at the bottom.

red nodes are interior and yellow represent branch heads

There is a toggle to switch between scrolling (transform) and selecting nodes
(picking).

When in picking mode the details of the commit or tag commit will show up in the
details panel below the graph itself.

The full comment is shown and the git-svn-id if it exists in the commment can be
used to provide better clarity.