KSENROLL-13873 add a simpified graph view
Instead of showing all of the nodes concentrate on branches and commits that
have multiple parents.

In simplified mode consolidate the commits that only have a single parent.

The Graph itself can also provide a count for the number of commits that point
at a specific parent.

For now just show the number of commits pointed at a comment when it has been
selected.