KSENROLL-13104 Create tool to remove .mpx files from project history
Removing the .sql files is non-trivial and complicated but the .mpx files are
included in the published jars.

It should be possible to purge them from the repository.

Adding a new cleaner that will let us test how much space can be reclaimed if
we remove all of the .mpx files.