KSENROLL-11841 Create branchName field in ExternalModuleInfo
Instead of always getting the branch name from a conversion of the branch path
just remember it.

Create a way to generate the string using the List<ExternalModuleInfo> directly.

In some fusion-maven-plugin use cases we want to update to the latest object id
on the branches given so do that first to change the externals list and then
use the new method to generate the String.

Also added a subTreePath field.  This is not used yet but will be in the future
to handle cases where the path given in the svn:external is really a branch
and then a subpath within the branch.

This will handle certain corner cases where the branch is created from a
copyfrom at a level that is not quite the same as in subversion.