KSENROLL-11841 Fix TextContentLength not defined but seen as defined bug
When the propContentLength == the contentLength we checked to see if the
textContentLength was defined (it is when there are real changes to a file)

But we initialized textContentLength to zero which was giving false positives.

Initialize to -1 so that we skip over non applicable file properties.

Adds a test case to verify it works.