KSENROLL-11841 Fix fusion data not updating bug and add r22 sample import test
The fusion data was only updating when there was an svn:externals property set
which is not correct.

The fusion data should be updated whenever there is a commit on the aggregate
trunk (or branch containing the fusion-maven-plugin.dat) and we notice that the
branch heads refered to in the fusion-maven-plugin.dat file have advanced.

Added a new revision 22 to the sample import test suite that tests this.

Originally it failed but with the changes contained here it passes.